refactoring + Matrix4

+++ src/Renderer/DefaultRenderer.cpp
@@ -1,16 +1,16 @@
-#include "../../include/Renderer/DefaultRenderer.h"
-#include "../../include/Light.h"
+#include "DefaultRenderer.h"
+#include "../Light/Light.h"
 #include <limits>
 #include <iostream>
-#define INF std::numeric_limits<double>::infinity()
+#define INF std::numeric_limits<float>::infinity()
 
-Color DefaultRenderer::getColor(Scene &scene, Ray &ray, double &x, double &y) {
+Color DefaultRenderer::getColor(Scene &scene, Ray &ray, float &x, float &y) {
 	Object* hitObject;
 	while (true) {
 		hitObject = 0;
-		double mint = INF;
+		float mint = INF;
 		for (int i = 0; i < scene.objects.size(); i++) {
-			double t = scene.objects[i]->hit(ray);
+			float t = scene.objects[i]->hit(ray);
 			if (!isnan(t) && t < mint && t > 0) {
 				mint = t;
 				hitObject = scene.objects[i];
@@ -27,7 +27,7 @@ Color DefaultRenderer::getColor(Scene &scene, Ray &ray, double &x, double &y) {
 bool DefaultRenderer::intersects(Scene &scene, Point3 &p, Vec3 &l) {
 	Ray ray(p, l);
 	for (int i = 0; i < scene.objects.size(); i++) {
-		double t = scene.objects[i]->hit(ray);
+		float t = scene.objects[i]->hit(ray);
 		if (!isnan(t) && t > 0.00001) {
 			return true;
 		}
